change chainId type and update sign method behavior

+++ besu/src/main/java/org/web3j/tx/PrivateTransactionManager.java
@@ -166,7 +166,7 @@ public abstract class PrivateTransactionManager extends TransactionManager {
 
     @Override
     public EthSendTransaction sendEIP1559Transaction(
-            BigInteger chainId,
+            long chainId,
             BigInteger maxPriorityFeePerGas,
             BigInteger maxFeePerGas,
             BigInteger gasLimit,
