+++ core/src/main/java/org/web3j/tx/ReadonlyTransactionManager.java
@@ -48,7 +48,7 @@ public class ReadonlyTransactionManager extends TransactionManager {
 
     @Override
     public EthSendTransaction sendEIP1559Transaction(
-            BigInteger chainId,
+            long chainId,
             BigInteger maxPriorityFeePerGas,
             BigInteger maxFeePerGas,
             BigInteger gasLimit,
